Reduce SDL dependency and warnings on Visual Studio
from 75 to 50

+++ src/engine/screen.cpp
@@ -1297,7 +1297,7 @@ namespace fheroes2
         Image::resize( width_, height_ );
 
         // To detect some UI artifacts by invalid code let's put all transform data into pixel skipping mode.
-        std::fill( transform(), transform() + width() * height(), 1 );
+        std::fill( transform(), transform() + width() * height(), static_cast<uint8_t>( 1 ) );
     }
 
     bool Display::isDefaultSize() const
