+++ src/fheroes2/game/game_highscores.cpp
@@ -383,7 +383,7 @@ void HGSData::ScoreRegistry( const std::string & p, const std::string & m, u32 r
 
     h.player = p;
     h.land = m;
-    h.localtime = std::time( nullptr );
+    h.localtime = static_cast<uint32_t>( std::time( nullptr ) );
     h.days = r;
     h.rating = s;
